Also delete self infering objects

+++ octopoes/octopoes/core/service.py
@@ -162,7 +162,11 @@ class OctopoesService:
 
     def _delete_ooi(self, reference: Reference, valid_time: datetime) -> None:
         referencing_origins = self.origin_repository.list_origins(valid_time, result=reference)
-        if not any(origin for origin in referencing_origins if origin.origin_type != OriginType.AFFIRMATION):
+        if not any(
+            origin
+            for origin in referencing_origins
+            if origin.origin_type not in [OriginType.AFFIRMATION, OriginType.INFERENCE] and origin.source == reference
+        ):
             self.ooi_repository.delete(reference, valid_time)
 
     def save_origin(
